Align better with WMA
Graph Construction & Representation:
https://reference.wolfram.com/language/guide/GraphConstructionAndRepresentation.html

Move eval stuff to pymathics.graph.eval
